Move RestControllerTrait to correct namespace
